move utils to the util test module
